+++ org.adempiere.asset/src/main/java/base/org/compiere/model/MAssetAddition.java
@@ -11,7 +11,7 @@ import org.compiere.FA.feature.UseLifeImpl;
 import org.compiere.process.DocAction;
 import org.compiere.process.DocumentEngine;
 import org.compiere.process.ProcessInfo;
-import org.compiere.process.ProjectClose;
+import org.compiere.project.process.ProjectClose;
 import org.compiere.util.CLogger;
 import org.compiere.util.DB;
 import org.compiere.util.Env;
