+++ org.adempiere.project/src/main/java/org/compiere/project/process/ProjectSetType.java
@@ -14,7 +14,7 @@
  * ComPiere, Inc., 2620 Augustine Dr. #245, Santa Clara, CA 95054, USA        *
  * or via info@compiere.org or http://www.compiere.org/license.html           *
  *****************************************************************************/
-package org.compiere.process;
+package org.compiere.project.process;
 
 import org.compiere.model.MProject;
 import org.compiere.model.MProjectType;
