+++ org.adempiere.project/src/main/java/org/eevolution/project/process/CopyFromProjectType.java
@@ -14,7 +14,7 @@
  * Created by victor.perez@e-evolution.com , www.e-evolution.com
  */
 
-package org.eevolution.process;
+package org.eevolution.project.process;
 
 import org.compiere.model.MProjectType;
 import org.compiere.model.MProjectTypePhase;
