+++ org.adempiere.project/src/main/java/org/eevolution/project/process/ImportProject.java
@@ -14,7 +14,7 @@
  * Created by victor.perez@e-evolution.com , www.e-evolution.com
  */
 
-package org.eevolution.process;
+package org.eevolution.project.process;
 
 import org.compiere.model.MActivity;
 import org.compiere.model.MBPGroup;
