+++ org.adempiere.project/src/main/java/org/eevolution/project/process/ImportProjectAbstract.java
@@ -14,7 +14,7 @@
  * Created by victor.perez@e-evolution.com , www.e-evolution.com
  */
 
-package org.eevolution.process;
+package org.eevolution.project.process;
 
 import org.compiere.process.SvrProcess;
